SOC mutants
PTW mutants are converted back to original and SOC mutants are generated.

+++ src/main/java/io/reactivex/rxjava3/core/Flowable.java
@@ -252,11 +252,7 @@ public abstract class Flowable<@NonNull T> implements Publisher<T> {
      */
     @CheckReturnValue
     public static int bufferSize() {
-        /**
-         * @pmt ptw#36
-         * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-         */
-        return (Integer.valueOf(BUFFER_SIZE)).intValue();
+        return BUFFER_SIZE;
     }
 
     /**
@@ -7750,11 +7746,7 @@ public abstract class Flowable<@NonNull T> implements Publisher<T> {
         Objects.requireNonNull(mapper, "mapper is null");
         ObjectHelper.verifyPositive(prefetch, "prefetch");
         Objects.requireNonNull(scheduler, "scheduler is null");
-        /**
-         * @pmt ptw#29
-         * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-         */
-        return RxJavaPlugins.onAssembly(new FlowableConcatMapScheduler<>(this, mapper, (Integer.valueOf(prefetch)).intValue(), ErrorMode.IMMEDIATE, scheduler));
+        return RxJavaPlugins.onAssembly(new FlowableConcatMapScheduler<>(this, mapper, prefetch, ErrorMode.IMMEDIATE, scheduler));
     }
 
     /**
@@ -16558,11 +16550,7 @@ public abstract class Flowable<@NonNull T> implements Publisher<T> {
     @SchedulerSupport(SchedulerSupport.CUSTOM)
     @NonNull
     public final Flowable<T> take(long time, @NonNull TimeUnit unit, @NonNull Scheduler scheduler) {
-        /**
-         * @pmt ptw#18
-         * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-         */
-        return takeUntil(timer((Long.valueOf(time)).longValue(), unit, scheduler));
+        return takeUntil(timer(time, unit, scheduler));
     }
 
     /**
@@ -16892,11 +16880,7 @@ public abstract class Flowable<@NonNull T> implements Publisher<T> {
     @SchedulerSupport(SchedulerSupport.CUSTOM)
     @NonNull
     public final Flowable<T> takeLast(long time, @NonNull TimeUnit unit, @NonNull Scheduler scheduler, boolean delayError, int bufferSize) {
-        /**
-         * @pmt ptw#17
-         * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-         */
-        return takeLast(Long.MAX_VALUE, (Long.valueOf(time)).longValue(), unit, scheduler, delayError, bufferSize);
+        return takeLast(Long.MAX_VALUE, time, unit, scheduler, delayError, bufferSize);
     }
 
     /**
