+++ src/main/java/io/reactivex/rxjava3/internal/operators/maybe/MaybeConcatArrayDelayError.java
@@ -116,11 +116,7 @@ public final class MaybeConcatArrayDelayError<T> extends Flowable<T> {
 
         @SuppressWarnings("unchecked")
         void drain() {
-            /**
-             * @pmt ptw#33
-             * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-             */
-            if (getAndIncrement() != (Integer.valueOf(0)).intValue()) {
+            if (getAndIncrement() != 0) {
                 return;
             }
