+++ src/main/java/io/reactivex/rxjava3/internal/util/BackpressureHelper.java
@@ -87,11 +87,7 @@ public final class BackpressureHelper {
      */
     public static long addCancel(@NonNull AtomicLong requested, long n) {
         for (;;) {
-            /**
-             * @pmt ptw#14
-             * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-             */
-            long r = (Long.valueOf(requested.get())).longValue();
+            long r = requested.get();
             if (r == Long.MIN_VALUE) {
                 return Long.MIN_VALUE;
             }
@@ -113,11 +109,7 @@ public final class BackpressureHelper {
      */
     public static long produced(@NonNull AtomicLong requested, long n) {
         for (;;) {
-            /**
-             * @pmt ptw#11
-             * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-             */
-            long current = (Long.valueOf(requested.get())).longValue();
+            long current = requested.get();
             if (current == Long.MAX_VALUE) {
                 return Long.MAX_VALUE;
             }
@@ -141,11 +133,7 @@ public final class BackpressureHelper {
      */
     public static long producedCancel(@NonNull AtomicLong requested, long n) {
         for (;;) {
-            /**
-             * @pmt ptw#15
-             * Generated artificial performance mutant for IEEE paper. DOI: 10.1109/TSE.2022.3188005
-             */
-            long current =  (Long.valueOf(requested.get())).longValue();
+            long current = requested.get();
             if (current == Long.MIN_VALUE) {
                 return Long.MIN_VALUE;
             }
